xiaohongshu/requirements.txt
# FastAPI Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
jinja2>=3.1.4

# OpenAI and LLM
openai>=1.12.0

# MCP (Model Context Protocol)
mcp>=0.9.0

# Environment variables
python-dotenv>=1.0.0

# Async support
aiohttp>=3.9.3

# HTTP requests
requests>=2.31.0
httpx>=0.27.0

# PDF Processing & ArXiv
pdf2image>=1.17.0
arxiv>=2.1.0
PyMuPDF>=1.23.8

# Other utilities
pydantic>=2.9.2
schedule==1.2.1
